+++ feature/card/src/main/java/com/itschristmas/card/cardeditor/ui/dialog/ObjectDeleteConfirmDialog.kt
@@ -1,8 +1,8 @@
-package com.itschristmas.card.cardeditor.dialog
+package com.itschristmas.card.cardeditor.ui.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
-import com.itschristmas.card.cardeditor.common.BaseDialog
+import com.itschristmas.card.cardeditor.ui.common.BaseDialog
 import com.itschristmas.card.R
 
 @Composable
